fix: createPurchase mutation
This commit fixes the "'PurchaseModel' instance needs to have a primary key value before this relationship can be used." error.

The error was caused by the fact that the purchase was not saved when the total was calculated on the PurchaseModel save method.
The solution was to create a recalculate_total method on the PurchaseModel that calculates the total and saves the purchase.
This method is called on a post_save signal of the PurchaseItemModel.
Thi way the total is always calculated when a PurchaseItemModel is updated or created.